Remove all dcgmproftester
`dcgmproftester` has diffetent versions and needs libcuda which is
not currently shipped in the snap, so it does not work out of the
box. Moreover, some versions like the `dcgmproftester10` is not
present in the arm64 version that breaks the build.

This PR removes all dcgmproftester until we have a more clear way
on how to ship this service and keep the support for multi-arch

Closes: #18